create SensorFactory
