+++ src/main/java/org/carlmontrobotics/lib199/MotorControllerFactory.java
@@ -7,29 +7,24 @@
 
 package org.carlmontrobotics.lib199;
 
+import org.carlmontrobotics.MotorConfig;
+import org.carlmontrobotics.lib199.sim.MockSparkMax;
+import org.carlmontrobotics.lib199.sim.MockTalonSRX;
+import org.carlmontrobotics.lib199.sim.MockVictorSPX;
+
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 import com.ctre.phoenix.motorcontrol.can.WPI_VictorSPX;
-import com.ctre.phoenix.sensors.CANCoder;
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMax.ExternalFollower;
 import com.revrobotics.CANSparkMax.IdleMode;
 import com.revrobotics.CANSparkMaxLowLevel;
 import com.revrobotics.SparkMaxPIDController;
 
-import org.carlmontrobotics.MotorConfig;
-import org.carlmontrobotics.lib199.sim.MockSparkMax;
-import org.carlmontrobotics.lib199.sim.MockTalonSRX;
-import org.carlmontrobotics.lib199.sim.MockVictorSPX;
-import org.carlmontrobotics.lib199.sim.MockedCANCoder;
-
-import edu.wpi.first.cameraserver.CameraServer;
-import edu.wpi.first.cscore.UsbCamera;
-import edu.wpi.first.cscore.VideoSource.ConnectionStrategy;
 import edu.wpi.first.wpilibj.RobotBase;
 
 /**
- * Add your docs here.
+ * A class containing methods to create and configure motor controllers.
  */
 public class MotorControllerFactory {
   public static WPI_VictorSPX createVictor(int port) {
@@ -150,37 +145,4 @@ public class MotorControllerFactory {
     return spark;
   }
 
-  public static CANCoder createCANCoder(int port) {
-    CANCoder canCoder = new CANCoder(port);
-    if(RobotBase.isSimulation()) new MockedCANCoder(canCoder);
-    return canCoder;
-  }
-
-  /**
-   * Configures a USB Camera.
-   * See {@link CameraServer#startAutomaticCapture} for more details.
-   * This MUST be called AFTER AHRS initialization or the code will be unable to connect to the gyro.
-   *
-   * @return The configured camera
-   */
-  public static UsbCamera configureCamera() {
-    UsbCamera camera = CameraServer.startAutomaticCapture();
-    camera.setConnectionStrategy(ConnectionStrategy.kKeepOpen);
-    CameraServer.getServer().setSource(camera);
-    return camera;
-  }
-
-  /**
-   * This method is equivalent to calling {@link #configureCamera()} {@code numCameras} times.
-   * The last camera will be set as the primary Camera feed.
-   * To change it, call {@code CameraServer.getServer().setSource()}.
-   *
-   * @param numCameras The number of cameras to configure
-   * @return The configured cameras.
-   */
-  public static UsbCamera[] configureCameras(int numCameras) {
-    UsbCamera[] cameras = new UsbCamera[numCameras];
-    for(int i = 0; i < numCameras; i++) cameras[i] = configureCamera();
-    return cameras;
-  }
 }
